+++ iosApp/iosApp/Features/Profile/PassportScreenView.swift
@@ -17,9 +17,9 @@ struct PassportScreenView: View {
             displayName: "Minh Nguyen",
             avatarUrl: nil,
             bio: "Capturing life memory stamps",
-            stampsCreatedCount: 14,
-            stampsCollectedCount: 38,
-            placesVisitedCount: 9
+            stampsCreatedCount: Int32(14),
+            stampsCollectedCount: Int32(38),
+            placesVisitedCount: Int32(9)
         )
     }
